pyproject.toml
[tool.poetry]
name = "test-service"
version = "0.1.0"
description = ""
authors = ["last-ui <slastukhin_av@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.isort]
profile = "black"
line_length = 79
skip_glob = ["**/migrations/*.py"]
src_paths = ["backend"]

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
